shrink to fit
size: 4079
bench: 3424185

+++ src/search.cpp
@@ -20,7 +20,7 @@ struct Searcher {
     uint64_t nodes;
     double abort_time;
     int16_t evals[256];
-    int64_t corr_hist[2][CORR_HIST_SIZE];
+    int16_t corr_hist[2][CORR_HIST_SIZE];
     HTable history;
     HTable conthist[14][SQUARE_SPAN];
     HTable *conthist_stack[256];
